Add Context Recorder interface + passthrough
Add Context Recorder interface to allow parser to record it's state as
it attempts to parse an input. This is put in place to allow for syntax
error reporting. The interface is made abstract and public to also allow
for custom tracers to be implemented.

Also implement a large refactor to simplify generated code. Lot's of
functionality pulled out into generic functions.