Add special @ match for position marking
This commit adds a new match called @. This is a special match that
doesn't consume any input and returns the current position of the
parser.

Example:

rule := 'a*' a_end=@ 'b*' b_end=@ 'c*'

This rule matches the set of strings 'a*b*c*', but in the created AST
node we get a properties called "a_end" and "b_end". In these properties
we store the PosInfo objects denoting where the sequence of a's ends,
and the sequence of b's ends respectively.

This commit also makes PosInfo interface readonly so that it doesn't
change after being stored.