Add flag to disable autogen comment
Resolves issue #40

The ability to disable this comment is needed in case "*/" appears  in
the grammar definition.

There is no equivalent of string escapes for comments, so its either
this or some compromise where we automatically break the */ sequence.
This seems to me it would break the whole point of the included grammar
comment so I choose this route.